refactor: implement SettingsModalShell for improved Dashlet settings management
- Introduced a new SettingsModalShell component to streamline the settings modal for Dashlet components, enhancing code organization and reusability.
- Refactored DashletSettings to utilize the new modal shell, consolidating visualization and data tabs for better user experience.
- Removed unused state management and imports, improving overall code clarity and maintainability.
